% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{validate_all_paths_out}
\alias{validate_all_paths_out}
\title{Validate the output of all_paths_fun}
\usage{
validate_all_paths_out(x, required_cols = NULL)
}
\arguments{
\item{x}{Object to validate.}

\item{required_cols}{Optional character vector of columns that must be present
in \code{x$nodes}.}
}
\value{
Invisible \code{TRUE} on success; throws an error otherwise.
}
\description{
Validate the output of all_paths_fun
}
\keyword{internal}
